requirements.txt
langchain_openai
langgraph
langchain_community
python-dotenv
langchain_experimental
pinecone
pinecone-plugin-assistant
pandas
nltk
seaborn
wordcloud
matplotlib
xlrd
flask
langchain
openai
gunicorn
typing-extensions>=4.0.0
waitress
